samples: driver: WM8904 config
- Project config updated
- Board overlays added:
	alif_b1_dk_ab1c1f4m51820hh0_rtss_he.overlay
	alif_b1_dk_ab1c1f4m51820ph0_rtss_he.overlay
- Compile warning fix

+++ samples/drivers/wm8904/src/main.c
@@ -49,7 +49,7 @@ static void tx_data_populate(void)
 		int16_t *right = &tx_data_block[i * I2S_CHANNEL_COUNT + 1];
 
 		int16_t sample = SIGNAL_AMPLITUDE *
-				 sin((2.0f * PI * SIGNAL_FREQUENCY * i) / I2S_SAMPLE_FREQUENCY);
+				 sinf((2.0f * PI * SIGNAL_FREQUENCY * i) / I2S_SAMPLE_FREQUENCY);
 		*left = sample;
 		*right = sample;
 	}
